poprawiona klasa z testami ShapeCollectorTestSuite

+++ kodilla-testing/src/main/java/com/kodilla/testing/shape/Shape.java
@@ -2,5 +2,5 @@ package com.kodilla.testing.shape;
 
 public interface Shape {
     String getShapeName();
-    int getField();
+    double getField();
 }
